V411-033 More precise message on call to subprogram out of SPARK
When a called subprogram is out of SPARK due to the SPARK_Mode of its
body, which is inherited by the subprogram spec when there is no
separate spec in the source, the solution is usually to separate the
spec from the body, so that it can be marked in SPARK. Improve the
error message in that case to help users.

[changelog]
* src/spark/spark_definition-violations.adb (Mark_Violation): Add
Cont_Msg parameter for an optional continuing message.
* src/spark/spark_definition-violations.ads (Mark_Violation): Change
profile.
* src/spark/spark_definition.adb (Mark_Call): Add special case.
* testsuite/gnatprove/tests/V411-033__spark_mode_off/main.adb
* testsuite/gnatprove/tests/V411-033__spark_mode_off/test.out
Add test.
* testsuite/gnatprove/tests/N710-015__spark_off/test.out
* testsuite/gnatprove/tests/O807-012__refined_post_not_in_spark/test.out
* testsuite/gnatprove/tests/R110-019__private_def_expr/test.out
Update output.

Change-Id: I2cf6b9667a3ead4af3c7125ad6c9628e0c8e77c6