Decrease CE_RAC stack depth
In counterexample runtime assertion checking (CE_RAC) we interpret user
code with a stack depth limit for recursive calls. This limit must match
a realistic limit for execution of recursive interpretation in gnat2why
itself.

The previous limit of 450 was too large and caused stack overflow,
in particular, when gnat2why was compiled with a very recent GNAT.
A new arbitrary limit of 100 seems to work fine.

* src/why/gnat2why-error_messages.adb
(Small_Step_Rac): Reduce stack limit; tune comment to avoid repeating
the exact value, which could easily get out-of-sync with the code.

no-tn-check

Change-Id: Id799cef586268586ae00fbb58d31e0272bccbcf9