V412-051 remove check for non-overlay address clauses
We used to check if the object type of an object declaration with an
address clause if suitable as the target of an unchecked conversion.
Now we do this check only if the address clause is an overlaying
one, of the form

  with Address => Obj'Address

for some other object Obj.

* gnat2why-expr.adb
(Transform_Decl): check if object type is suitable as a target for
  UC only if this address clause is an overlaying one.

Change-Id: I26cda8b0f540f2d20b11b7e9bce2e563966f96eb