V325-031 Update parsing of Why3 JSON output
Following the changes in the JSON output produced by Why3, it is
necessary to adapt the parsing on the SPARK side.

* src/common/vc_kinds.adb
(From_JSON): The "cntexmp" field is no more a mapping of filenames to
models but a list of records with two fields: "filename" and "model".
(From_JSON): Each "model" is no more a mapping of lines (either a line
number or "vc_line") to model elements for this line, but a list of
records with 3 fields: "loc", "is_vc_line" and "model_elements".
(From_JSON): The field "model_elements" contains a (possibly empty)
list of values.

* testsuite/gnatprove/tests/R420-029__parsing_mlw/test.out
Minor change in the JSON output.

Change-Id: I14d401c26e6111386c3677d607b3884c56feb3b7
Depends-On: I29db17a0060919e954374e8f67e4af106f09fcff