V503-011 accept conversions from integers or addresses to pointers
Converting from an integer type or System.Address to an
access-to-object type is considered too useful to be rejected. We
accept it with some warnings stating that the new pointer is assumed
to be initialized and without any aliases. It is similar to what we
do for overlays, but volatile properties and pragma Import cannot be
used to silence the warnings here, so overlays should still be the
preferred way of handling such cases if possible.
We also reject all kind of unchecked conversions to types with access
subcomponents, even those which are not deep. Indeed, as we do not
model pointer addresses, we cannot support them.

* src/spark/spark_definition.adb
(Mark_Subprogram_Entity): Accept unchecked conversions from integer
 types or System.Address to general access-to-object types. Make the
 error message more precise when the unchecked conversion is rejected.
(Mark_Call): Emit warnings on accepted calls to unchecked conversion
 if they return access types.

* src/spark/spark_util-types.adb
(Traverse_Access_Parts): Generic function which checks whether a type
 contains at least one access subcomponent on which a given test
 passes.
(Contains_Allocated_Parts, Is_Deep): Refactor to use
 Traverse_Access_Parts.
(Contains_Access_Subcomponents): New function which returns True if a
 type has any kind of access subcomponents.
(Traverse_Access_Parts): Return True on System.Address.

* tests/U622-049__deep_tagged_root/test.out
Sharing the code between Is_Deep and Contains_Allocated_Parts fixes a
bug in the detection of memory leaks which was introduced when
Contains_Allocated_Parts was not updated after deep subcomponents were
allowed in root tagged types.

* tests/P513-017__explicit_dereference/test.out
Previously rejected conversion is accepted. No warnings are emitted as
the conversion function is not called.

* tests/T619-025__private_unchecked/test.out
Conversion from a type with access-to-constant subcomponents is now
rejected.

* tests/U209-014__access_address/test.out
Changes in error messages.

* tests/V503-011__unchecked_conversion_access/
New test.

Change-Id: Id66935b7669c4d0a72ea04bd6b66844c501845de
Depends-On: I03cbf93f881cb3538c0aa01e3100c1cd4a2671f6