V422-015 take others value into account on array element assignment
When assigning a value to an element in an array Value_Type, an entry
in the Array_Values map was systematically created or updated, no matter
if the "others" value was sufficient. Now an element is entered in the
map or updated only when necessary.

The test Q912-021__codepeer_by_example timed out because for each
iteration of a loop, the very large array was getting bigger, along with
the time needed to copy that value (when returning from RAC_Statement
for example).

* src/counterexamples/ce_rac.adb
(RAC_Statement): Only add an entry in the Array_Values map if necessary

* tests/Q912-021__codepeer_by_example/test.out
Update test output

* tests/Q912-021__codepeer_by_example/test.yaml
Enable couterexample checking

Change-Id: I27a6b0d757beb776667e80ce4b539c921a9273d6