V506-021 fix translation of Iterable on subtypes
We use the base type for the loop parameters to account for empty
ranges. We should not do the same thing for iterators as First and
Next always return a cursor in the cursor type.

* src/why/gnat2why-expr-loops.adb
(Transform_Loop_Statement) Do no go to the base type for the loop
parameter of iterators. Also assume the dynamic property of the
parameter in the loop.

* tests/V506-021__infinite_sequences
File test.

Change-Id: I90949c1770efc3e2ab8b990300b39793caff379d