U908-033 add example with pointer with aliasing in separate memories
This is a variation of the hidden_pointers library where it is
possible to use different memory regions. The separation of the memory
regions is ensured using ownership. It makes it possible to define
data-structure with cycles or sharing in a single structure while
retaining non-aliasing between different structures. It is also possible
to transfer a pointer from a memory region to another manually. This
cannot be recognized as ghost code currently as the memory cannot be
ghost. On a side note, this one allows to prove absence of memory leaks
on the hidden pointers when the memory itself goes out of scope.

* tests/U908-033__hidden_pointers_separate_mem
File test.

Change-Id: I055ece3ae5f3b3d9edfe35ffa908d4905705efcd