QC04-054 Clarify use of state abstraction in Initializes contracts
In the case of a package spec in the private part of an enclosing
package, the package spec may have visibility over components of
abstract states while the refinement has not been seen yet.
In that case, the same rule as for Global contracts should be
enforced: the corresponding abstract state should be used.

Change-Id: Id085b02eb280450fcb8bf81e5b19779520a64e90