V517-002 remove disallowed case of X'Access.all in potential aliases
SPARK RM 4.1.4(4) states that a reference to the Access attribute
whose type is an anonymous access-to-object type or a named
access-to-variable type shall occur directly inside a stand-alone
object declaration, an assignment, or a return statement. Therefore, it
is not necessary to consider the case of a dereference whose prefix
contains a reference to the Access attribute in the definition of
potential aliases.

* docs/lrm/source/declarations-and-types.rst
Remove disallowed case in the definition of potential aliases.

Change-Id: I359121b908ee1df8400414cf4325e57206146fbd