Fix missing check when borrower exits its scope
When a borrower exits its scope, it should not have any reachable
parts in the Moved state. This is now implemented in the borrow-checker.
A corresponding rule will be separately added to the SPARK RM as part
of the rework of the rules regarding access types.

[changelog]
* src/why/gnat2why-borrow_checker.adb (Check_End_Of_Scope): New
procedure to check the permission of borrowers at the end of their
scope.
(Check_End_Of_Scopes): New procedure to check the permission of
borrowers in all scopes in a given range.
(Loop_Statement_Of_Exit): New function to return the loop node
targetted by an exit statement.
(Perm_Error_Borrow_End): New procedure to issue an error when permission
for a borrower at the end of its scope is not Read_Write.
(Check_Callable_Body, Check_Statement): Add call to Check_End_Of_Scope
and Check_End_Of_Scopes.
* testsuite/gnatprove/tests/V531-016__borrow/hole.adb
* testsuite/gnatprove/tests/V531-016__borrow/test.out
* testsuite/gnatprove/tests/V531-016__borrow/top.adb
Add test.

Change-Id: I33ab5a84a2c7d5a55593573b9b3fdb56e1834a5d
TN: V531-016