Avoid dummy conjuncts in generated Why AST API
In generated API for Why AST we had:

     function Type_Id_Children_Valid
       (Id : W_Type_Unchecked_Id)
       return Boolean is
       (True
       and then
         Id_Cache_Valid
           (Get_Node (Id).K1_Name)
       and then
         True
       and then
         True);

and CodePeer rightly complained:

  medium warning: same_logic (gnatcheck): same expression found

Actually, it is a CodePeer bug that this unit got now analysed, but it
seems more robust to simply avoid these repeated expressions and not
risk similar checks for other checkers that we might run in the future.

* src/why/xgen/xtree_children_checks.adb
(Print_Children_Check_Expression): Don't print unnecessary True
conjuncts.

no-tn-check

Change-Id: I74c20836cdfa26fdf4f37f77359de7d3b74d5af1