V609-053 correct handling of moves through 'Access
We consider when moving P'Access that only the subcomponents of P
which are deep are moved. For P.all'Access it is fine as memory leaks
are only checked on pool-specific access types, and the target of the
move is necessarilly a general access type.

* src/why/gnat2why-expr.adb
(Can_Be_Moved): Return False for a reference to 'Access.
(Collect_Moved_Objects): Collect objects in the prefix of 'Access.

* tests/V609-053__access_attribute
New test.

Change-Id: I6bc15723f060291514ac9b4c66eba6cdb56277fa