V422-022 Add unbounded and indefinite formal vectors
Before this patch, the only formal vector were bounded and definite.
This mean that it is necessary to provide their maximum length or
capacity at instantiation and that they can only be used with definite
element type.
The formal vectors added by this patch are unbounded and indefinite.
Their length grows dynamically until Count_Type'Last. This makes them
easier to use but requires the use of dynamic allocation and controlled
types.
This patch also adds a test for the new unit.

* include/spark-containers-formal-unbounded_vectors.adb
* include/spark-containers-formal-unbounded_vectors.ads
The unbouded vectors' implementation files.

* tests/V422-022__formal_unbounded_vectors/test.gpr
The project file.

* tests/V422-022__formal_unbounded_vectors/test.adb
The main test file.

* tests/V422-022__formal_unbounded_vectors/test_vectors.adb
* tests/V422-022__formal_unbounded_vectors/test_vectors.ads
Implements the tests.

* tests/V422-022__formal_unbounded_vectors/test.py
The test parameter file.

* tests/V422-022__formal_unbounded_vectors/test.out
The test output.

Change-Id: I76b9950f60286cc16787769f1ea26e976ca56bde