V601-015 use fuzzer when no prover counterexample
* src/why/gnat2why-error_messages.adb
* src/why/gnat2why-error_messages.ads
(Handle_Result): Change condition for using the fuzzer. Consume more
fuel. Add Print_Fuzzing variable to avoid printing a counterexample when
the fuzzer was interrupted before finding a good counterexample.
(Emit_Proof_Result): Add Print_Fuzzing parameter.
(Compute_Cannot_Prove_Message): If the fuzzer is responsible for the CE
verification, don't specify which bound could not be proved in the case
of an overflow check. The message is decided based upon the result of
the giant-step RAC which might not correspond to that of the fuzzer.

* src/counterexamples/ce_rac.adb
* src/counterexamples/ce_rac.ads
(RAC_Attribute_Reference): Small-step RAC crashes when 'Length attribute
used on type. Add call to RAC_Unsupported to avoid crash.
(Check_Fuel_Decrease): Add parameter to choose amount of fuel to
consume.

* src/flow/flow_error_messages.adb
* src/flow/flow_error_messages.ads
(Error_Msg_Proof): Values generated by the fuzzer were not printed
because of a condition on the existence of the CE which was entered
first. Take into accout the Fuzzing_Used parameter to avoid entering if
the CE was not used. Don't call Get_Environment_One_Liner if the fuzzer
did not have time to find a proper counterexample. Add Print_Fuzzing
parameter.

* tests/JC17-017__pre_guarded/test.out
* tests/K407-022__arith/test.out
* tests/LB27-011__overflow_config/test.out
* tests/LB27-011__overflow_pragma/test.out
* tests/LB27-011__overflow_switch/test.out
* tests/M115-016__reatim/test.out
* tests/M321-015__operator_effects/test.out
* tests/M523-040__global_v2_tests/test.out
* tests/N122-016__switches/test.out
* tests/N804-036__bad_prover/test.out
* tests/O512-022__vscomp2014_partition_functional/test.out
* tests/R723-013__memlimit/test.out
* tests/S327-015__itp_desired_tasks/test.out
* tests/S327-020__loop_inv/test.out
* tests/TU__refined_post_illegal/test.out
* tests/riposte__arithmetic/test.out
* tests/spark_book/test.out
Update test outputs

Change-Id: Iea62eca6ac1c9c2098eaa54c84589568ce3325f4