V422-022 Add new unbounded and indefinite formal ordered set
Before this patch, the only formal ordered sets were bounded and
definite. This means that it is necessary to provide their maximum
length or capacity at instantiation and that they can only be used with
definite element types.
The formal ordered Sets added by this patch are unbounded and
indefinite. Their length grows dynamically until Count_Type'Last.
This makes them easier to use but requires the use of dynamic
allocation and controlled types.
This patch also adds a test for the new unit.

* include/spark-containers-formal-unbounded_ordered_sets.adb
* include/spark-containers-formal-unbounded_ordered_sets.ads
The multisets' implementation files.

* tests/V422-022__formal_unbounded_ordered_sets/test.gpr
The project file.

* tests/V422-022__formal_unbounded_ordered_sets/test.adb
The main test file.

* tests/V422-022__formal_unbounded_ordered_sets/test_set.adb
* tests/V422-022__formal_unbounded_ordered_sets/test_set.ads
Implements the tests.

* tests/V422-022__formal_unbounded_ordered_sets/test.py
The test parameter file.

* tests/V422-022__formal_unbounded_ordered_sets/test.out
The test output.

Change-Id: I55b7fae981195141ede8fc80603c4ba647d2df9c