V622-008 fix classification of large tests
The test finder was returning too early when identifying a test as
an UG or RAC test. As a result, these tests were always run even
when only large tests were requested. The fix is to move the checks
for UG or RAC tests a bit later in the test finder code.

* run-tests
(probe): first check if test is large, only then check if UG or RAC

Change-Id: Ib6ff3f5c5e75c28c68079e1a8a7f8e9e38a37b13