Update the proof of Int_Division for the new axiom on mod
With the introduction of axiom Mod_Unique to have the two definitions of
mod be seen as equal, the Coq realization of Int_Division needs to be
updated. Also change the axiom so that the equality of the two
definitions of mod is only asserted for non-zero divisor (where mod is
not defined in SPARK, and could be defined in Coq depending on the
definition used, so better avoid defining it here).

[changelog]
* share/spark/theories/_gnatprove_standard.mlw (Mod_Unique): Add
premisse to the axiom to exclude zero divisor.
* ../gnatprove/tests/O702-009__coq_realizations/realize/Int_Division.v
(Mod_Unique): Prove the lemma.
* testsuite/gnatprove/tests/R427-016__check_on_param_call/test.out
* testsuite/gnatprove/tests/TA27-008__iterator_filter/test.out
Update outputs.
* testsuite/gnatprove/tests/prime_numbers/proof/sessions
Update session files.

Change-Id: I3aa9b670663d65e9b5931be22802e76fa982ebfe
TN: V617-028