V422-022 Add new unbounded and indefinite formal ordered maps
Before this patch, the only formal ordered maps were bounded and
definite. This means that it is necessary to provide their maximum
length or capacity at instantiation and that they can only be used
with definite element and key types.
The formal maps added by this patch are unbounded and indefinite.
Their length grows dynamically until Count_Type'Last. This makes them
easier to use but requires the use of dynamic allocation and controlled
types.
This patch also adds a test for this unit.

* include/spark-containers-formal-unbounded_ordered_maps.ads
* include/spark-containers-formal-unbounded_ordered_maps.add
Implementation files of the formal unbounded indefinite ordered maps.

* tests/V422-022__formal_unbounded_ordered_maps/test.gpr
The project file.

* tests/V422-022__formal_unbounded_ordered_maps/test.adb
The main test file.

* tests/V422-022__formal_unbounded_ordered_maps/test_map.adb
* tests/V422-022__formal_unbounded_ordered_maps/test_map.ads
Implements most of the tests.

* tests/V422-022__formal_unbounded_ordered_maps/pragmas.adc
Disable the check false, used as unsoundness detector, at run time.

* tests/V422-022__formal_unbounded_ordered_maps/test.py
The test parameter file.

* tests/V422-022__formal_unbounded_ordered_maps/test.out
The test output.

Change-Id: Ifac6e840be72cb1fce1edfb0b485e9ded61eda02