V712-038 use big integers in SPARK.Higher_Order.Fold.Sum
Replace the signed integer parameter type Element_Out in the generic
Sum and Sum_2 packages by a private type with Zero and Add functions.
Use conversion to big integers as a model of the summation.

* include/spark-higher_order-fold.ads
(Sum, Sum_2): Use a private type and user-provided Add and Zero
 functions for the return type of the Value function. The proofs are
 simpler as we have as a precondition that no overflows can occur in
 the summation. Users will need to prove this separately on their
 instances.

* tests/R509-011__higher_order_proof
Update the proofs of the Sum and Sum_2 lemmas. Also prove the absence
of overflows in the case of adding small integers as was done before.

* tests/R509-011__higher_order
Adapt the test to new signature of the generic Sum and Sum_2 packages.
Simplify the test and decrease the step limit so the test runs in a
reasonable amount of time.

Change-Id: I63aa4bf0fc3ba4861bfdfbe3b0ba5c2e5af98353