V714-006 introduce separate modules for DIC and type invariants
The default assumption and type invariant predicates used to be
declared in the type completion module. Due to how inclusion of
modules works, this caused the axiom modules of entities called in
these predicates to be included in the VC module of all entities
using this type. In particular, if a function is used in the DIC or
type invariant of the type, its axiom module will be withed in its
VC module, which is unsound. We now store these predicates in separate
modules which will only be withed if an object of the type is
effectively initialized by default or the type invariant is effectively
checked.

* src/why/gnat2why-types.adb
(Generate_Type_Completion): Generate the default assumption and type
 invariant predicates in their own module.

* src/why/why-atree-modules.adb
(DIC_Modules, Invariant_Modules): New maps storing these extra modules.
(E_DIC_Module, E_Invariant_Module): Functions returning the extra
 modules for the type invariant and default assumption predicates.
(Insert_Type_Symbols): Locate the default assumption and type invariant
 predicates in the appropriate modules.

* tests/V714-006__multisets
Enable test.

* tests/K624-029__amortize_queue/test.out
* tests/L719-025__ordered_maps_equal/test.out
* tests/MB14-009__for_loops_iterable/test.out
* tests/S715-026__dic_loop/test.out
* tests/U908-033__hidden_pointers/test.out
* tests/U908-033__hidden_pointers_separate_mem/test.out
* tests/V318-027__counterex_loop_entry/test.out
* tests/V422-022__formal_indefinite_doubly_linked_list/test.out
* tests/V714-006__multisets/multiset.ads
* tests/V714-006__multisets/test.opt
* tests/V714-006__multisets/test.out
* tests/ug__long__invariants/test.out
* tests/O512-022__vscomp2014_partition_functional/test.out
* tests/U413-014__container_references/test.out
* tests/O512-022__robot_navigation_snd/test.out
* tests/PC14-014__functional_annotations/test.out
Small diffs in provability due to provers instability.

Change-Id: I447a85552c7255fa13fb3dcc5c52ebe3e0e44d50