V714-006 use integer value of tags in compatibility axioms
To avoid unecessary uses of tagged type modules, use the integer value
of the tag in the compatibility axioms for a tagged hierarchy. This is
some preliminary work before changing the handling of module inclusion.

* src/utils/common_containers.ads
(Node_To_Int_Maps): Maps of nodes to integers.

* src/why/why-gen-records.adb
(Next_Tag): Removed.
(Tag_Representation): Module for the representation of the tag of
 tagged types as integers. Provides a Get_Tag function which returns
 the int value to be used for the tag of a tagged type.
(Create_Compatible_Tags_Theory): Use the int value of the tags for the
 compatibility axiom.

* src/why/gnat2why-tables.adb
(Init_Component_Info): Do not include subtypes in descendant set.

* src/why/gnat2why-subprograms.adb
(Generate_Dispatch_Compatibility_Axioms): Go to the base type for the
 diaptching type.

* tests/O901-014__formal_vectors/test.out
* tests/Q529-037__discriminant/test.out
* tests/N709-001__tagged_discr/test.out
Minor change in provability.

Change-Id: I66e02a4041d0d2c7828a01e4112d6f521de4395c