V726-023 improve violation messages related to limited with
We do not support using the limited view of a type coming from a
limited with yet. Before this patch, the violation was emitted on the
entity of the limited view. Unfortunately, as this entity has the
location of the complete view of the type, this results in messages
emitted inside the package which is limited with, and not in the unit
which uses it. This patch avoids marking limited views of types
altogether by rather rejecting the contsruct using the limited view.
This patch goes in the right direction if we want to support these
partial views at some point, as these views do not have the flags
set in the frontend like their full view, so they should probably
be by-passed in the analysis.

* src/spark/spark_definition-violations.ads
(Mark_Unsupported): Add an optional entity parameter and Root_Cause_Msg
 to be able to reference an entity in the error message.

* src/spark/spark_definition.adb
(Is_Incomplete_Type_From_Limited_With): Return true of the limited view
 of a type coming from a limited with.
(Reject_Incomplete_Type_From_Limited_With): Reject an entity which is
 using the limited view of a type.
(In_SPARK, Retysp_In_SPARK): Reject limited views without marking them.
(Mark_Parameter_Entity, Mark_Function_Specification, Mark_Type_Entity):
 Reject uses of limited views in subprogram parameters, return type of
 functions, and designated type of access types.

* src/spark/spark_util-types.adb
(Acts_As_Incomplete_Type): Explicitly reject limited views.

* tests/PA03-014__limited_with/test.out
* tests/V606-041__limited_with/test.out
Changes in violations.

* tests/V726-023__limited_with
New test.

Change-Id: If53bbbc716559f0081f587630c09e5cf7d257edd