Keep proof from using Refined_Global of expr funcs from other units
When proof peeks into expression functions from a WITH-ed units, it
might effectively peek into a (generated) Refined_Global that otherwise
should not be visible.

[changelog]
* src/flow/flow_utility.adb (Get_Proof_Globals): When getting Global of
expression function declared in a WITH-ed unit, use the "visibility
scope" of its spec, so that its Refined_Global does not become visible.

* testsuite/gnatprove/tests/V802-036__flow_expr_fun/
Enable; no output is a good output.

TN: V802-036
Change-Id: Icaa07eaa920bce65ddfc99ae73fe8f3fd3e75302