V804-026 add a test on Replace for formal ordered sets
Add a test to make sure that the error on contracts disappeared.

* tests/V804-026__bug_contracts_formal_ordered_set/test.adb
The Ada code for the test.
* tests/V804-026__bug_contracts_formal_ordered_set/test.gpr
The test project's main file.
* tests/V804-026__bug_contracts_formal_ordered_set/test.out
The current test outputs.
* tests/V804-026__bug_contracts_formal_ordered_set/test.py
The test configuration file.

* tests/T210-005__conts/test.out
Update the test output because of a line change introduced by
the patch of the postcondition of the Replace function.

* tests/L719-025__ordered_sets_key/test.out
Update the test output. The postcondition is no more proovable because
of the changes in the contract of Replace. This should be discuss in
a further ticket

Change-Id: I157a5bb7e50491999da876d3d8512cec16fa1262