V808-020 add test with null path
Add minimal reproducer with NULL as an actual parameter of a subprogram
call with no effects. This call needs to be wrapped in a subprogram with
some effects (e.g. a formal parameter of mode IN OUT), as otherwise the
analysis of statements with no effects is bypassed.

* testsuite/gnatprove/tests/V808-020__flow_null/
Add; disabled, because it causes a crash in flow.

Change-Id: Id154e159099fdacf0ffc1f4cae6689e4852a971d