V818-010 Add the functional Multisets
This patch adds functional Multisets to SPARK. Functional Multisets
are functional containers based on the functional Maps. Basicaly,
Multisets are sets which stores all together the elements and their
number of occurence.
This containers are useful to give an high level view on containers
which hold mutiple time the same element. This, for exemple, allows
to make sure that the overall content of a container is not changed
by a sorting procedure.

* include/spark-containers-functional-multisets.ads
* include/spark-containers-functional-multisets.adb
The two implementation files of the multiset.

* include/spark-containers-functional-multisets__light.ads
* include/spark-containers-functional-multisets__light.adb
Non-executable variants compatible with the light runtime.

* include/sparklib_light.gpr
Add light version of multisets.

* tests/V818-010__functional_multiset_proof/my_multiset.ads
* tests/V818-010__functional_multiset_proof/test.adb
* tests/V818-010__functional_multiset_proof/test.gpr
* tests/V818-010__functional_multiset_proof/test.out
* tests/V818-010__functional_multiset_proof/test.py
* tests/V818-010__functional_multiset_proof/types.ads
Implements the proof of the multisets.

* tests/V818-010__functional_multiset_test/natural_multisets.ads
* tests/V818-010__functional_multiset_test/pragmas.adc
* tests/V818-010__functional_multiset_test/test.adb
* tests/V818-010__functional_multiset_test/test.gpr
* tests/V818-010__functional_multiset_test/test.out
* tests/V818-010__functional_multiset_test/test.py
* tests/V818-010__functional_multiset_test/test_multiset.adb
* tests/V818-010__functional_multiset_test/test_multiset.ads
Implements the tests for the multisets.

* tests/V818-010__functional_multiset_proof/proof
The replay files used by V818-010__functional_multiset_proof.

* tests/V916-025__sparklib_light
Instantiate multisets.

Change-Id: I8100a8159f07861f7aa198d123be376b2bbb3291