V818-016 Avoid issuing warnings on intentional dead branch
Calling a function with a precondition of False is a way to indicate
that a branch inside a conditional expression is unreachable. This will
be checked by proof, so no warning should be issued about this branch
being unreachable when using --proof-warnings.

[changelog]
* src/spark/spark_util.adb (Is_Error_Signaling_Statement): Add special
case for functions with precondition of False.
* src/spark/spark_util.ads (Is_Error_Signaling_Statement): Update
comment.
* src/why/gnat2why-subprograms.adb (Generate_VCs_For_Subprogram):
Fix setting of boolean to determine when a precondition is always False,
which should be independent of whether a warning is issued on statically
False preconditions, as this is used later to decide on the generation
of a VC when --proof-warnings is used.
* /gnatprove/tests/V818-016__warning_pre_false/no_warn_unreachable.adb
* testsuite/gnatprove/tests/V818-016__warning_pre_false/test.out
* testsuite/gnatprove/tests/V818-016__warning_pre_false/test.yaml
Add test.

Change-Id: I1fc78844c728cc573f3fefb7d8378c482a3b3b5d