V906-009 Add a new test for the formal ordered sets
This patch adds a new test for the formal ordered sets. It is based
on PC14-014__formal_containers and aims at providing a better
coverage of the unit.
It uses the replay mode to reduce the test duration.

* tests/V906-009__formal_ordered_sets/pragmas.adc
* tests/V906-009__formal_ordered_sets/test.adb
* tests/V906-009__formal_ordered_sets/test.gpr
* tests/V906-009__formal_ordered_sets/test.out
* tests/V906-009__formal_ordered_sets/test.py
* tests/V906-009__formal_ordered_sets/test_set.adb
* tests/V906-009__formal_ordered_sets/test_set.ads
Implements and sets up the tests.

* tests/V906-009__formal_ordered_sets/proof
The directory which contain all the session files use by replay.

Change-Id: I477472f2f0b02aa23546f70c91063ef8e87fb153