V906-009 Add a new test for the formal hashed maps
This patch adds a new test for the formal hashed maps. It is based
on PC14-014__formal_containers and aims at providing a better
coverage of the unit.
It uses the replay mode to reduce the test duration.

* tests/V906-009__formal_hashed_maps/pragmas.adc
* tests/V906-009__formal_hashed_maps/test.adb
* tests/V906-009__formal_hashed_maps/test.gpr
* tests/V906-009__formal_hashed_maps/test.out
* tests/V906-009__formal_hashed_maps/test.py
* tests/V906-009__formal_hashed_maps/test_map.adb
* tests/V906-009__formal_hashed_maps/test_map.ads
Implements and sets up the tests.

* tests/V906-009__formal_hashed_maps/proof
The directory which contain all the session files use by replay.

Change-Id: I9574df94501e43425dd13ecda8b51c6bc851c865