V906-009 Add new tests for the formal list
This patch adds a more complete test for the formal doubly linked
list. A part of this test is based on PC14-014__formal_containers.

As the test is quite long, it uses the replay mode.

* V906-009__formal_doubly_linked_lists/test.adb
* V906-009__formal_doubly_linked_lists/test.gpr
* V906-009__formal_doubly_linked_lists/test.out
* V906-009__formal_doubly_linked_lists/test.py
* V906-009__formal_doubly_linked_lists/test_list.adb
* V906-009__formal_doubly_linked_lists/test_list.ads
Implements the new tests and set up the run procedure.

* V906-009__formal_doubly_linked_lists/proof
Session files for the replay mode.

Change-Id: I6f8d49d70f5f96334f02e6a78a69d17c7608c4d1