V901-002 Document miscellaneous warnings issued by GNATprove
In addition to the warnings issued by flow analysis and proof, also
document other warnings possibly issued by GNATprove, so that users
can decide on the severity of each kind of warnings.

[changelog]
* docs/ug/Makefile: Add generated file.
* docs/ug/en/source/how_to_view_gnatprove_output.rst: Add new table.
* docs/ug/table_generator/table_generator.adb: Produce new table.
* src/common/vc_kinds.adb (Description, Kind_Name): New functions
returning the description of a warning for the User's Guide.
* src/common/vc_kinds.ads (Misc_Warning_Kind): New type for misc
warnings.
(Warning_Message): New function with text of warning message (which
may includes wildcards for Error_Msg functions).
* src/spark/spark_definition-annotate.adb: Update to use warning kinds.
* src/spark/spark_definition.adb: Same.
* src/why/gnat2why-subprograms.adb: Same. Also rurn warnings on
annotation for proof into info messages, to be consistent with other
messages only related to ease of proof.
* src/why/why-gen-hardcoded.adb: Same.

Change-Id: I753bde59243cc1b0dd133588140d53d2a53c7ac9