V907-020 do not protect function axioms from the SPARK library
Special case functions from the SPARK library to not use guards to
protect their post axioms. Indeed, there should not be any function
with an unsound (implicit or explicit) postcondition in the SPARK
library.

* src/spark/spark_util.adb
(In_SPARK_Library_Unit): Return True if the source unit for N is in
 the SPARK library, ie. if the unit name starts with "spark-" and ends
 with either ".ads" or ".adb".

* src/why/gnat2why-util.adb
(Use_Guard_For_Function): Do not use guards for functions located in
 the SPARK library.

* gnatprove/testsuite/lib/python/test_support.py
Remove the temporary fix using --no-axiom-guard with for tests using
the spark library.

* tests/T727-043__mergesort
Update session test.

* tests/S117-008__pointer/test.out
* tests/T928-002__at_end_borrow/test.out
* tests/U129-014__sqrt_error_bounds/test.out
* tests/V708-031__functional_container_choose/test.out
* tests/adacore_u/test.out
* tests/sparksmt__lexer/test.out
* tests/sparksmt__lexer_termination/test.out
* tests/O512-022__robot_navigation_snd/test.out
* tests/spark_io/test.out
* tests/float_error_bounds/test.out
* tests/openETCS/test.out
Minor provability changes due to functions which are not coming from
spark library having a guard again.

Change-Id: I09db8a0b80c01786b8aa2f43e6ac0877ebb1542e