V422-022 Remove the formal indefinite vectors
Remove the formal indefinite vectors. Update the testes accordingly.
In the tests, the Indefinite_Vectors are replaced by the
Unbounded_Vectors as much as possible.

* include/spark-containers-formal-indefinite_vectors.adb
* include/spark-containers-formal-indefinite_vectors.ads
Remove the two implementation files.

* tests/N317-021__vectors
* tests/PC14-014__formal_containers
Remove the bounded part of the test and update the unbounded part.

* tests/O504-014__state_machine
* P414-029__equality_formal_vectors
* tests/P718-001__ubstr_bad
* tests/P718-001__ubstr_good
* tests/PC14-014__functional_annotations
* tests/sparksmt__lexer_termination
* tests/sparksmt__lexer
* tests/sparksmt__names
* tests/U413-014__container_references
Tests which used the unbounded version of the indefinite vectors
updated to use the Unbounded_Vectors.

* tests/O525-026__classwide
* tests/O901-014__formal_vectors
* tests/R920-050__indef_vector
* tests/U413-014__deep_element_type
* tests/UA15-028__discriminant_in_borrow
* UA25-034__implicit_access_conversion
Test which use bounded indefinite vectors are updated
to use Unbounded_Vectors.

* tests/PC14-014__formal_containers/proof
* tests/PC14-014__functional_annotations/proof
Update sessions files.

* tests/NB19-026__flow_formal_vectors
Remove the test because it was focusing on bounded vectors.

Change-Id: Ic22f6122ed4c084fff4b4f14107bd4ce24d2c292