VA06-022 add guards to array comparison axioms
So they do not imply that there exists a value of the representative
index type smaller than the first index of all arrays. This was not
correct for arrays indexed by modular types.

* share/spark/theories/ada__model.mlw
(Array_Comparison_Axiom): Add guards in defining axioms for array
 comparison Lt and Gt. Also introduce a lcp function which produces
 a witness used to replace the existentially quantified formulas in
 these axioms. This improves provability.

* tests/O702-009__coq_realizations
Realize the array comparison axioms.

* tests/VA06-022__modular_array_comparison
New test. Disable counterexample checking as it crashes for some
reason.

* tests/M806-034__array_comparison/test.out
* tests/O416-002__type_contracts/test.out
* tests/T630-007__relaxed_init/test.out
* tests/V422-022__formal_indefinite_ordered_maps/test.out
* tests/V422-022__formal_indefinite_ordered_sets/test.out
Minor diffs in provability and improvements due to the skolemization.

Change-Id: Iab8be6060d2a589306ff7dd218b1342dfe8bed73