VA05-041 Fix bad interaction between ownership and relaxed init
When a type subject to ownership is included in a record/array
type subject to relaxed initialization, the wrapper functions
to convert between the two possible views of the type did not
take the Is_Moved component into account. Now fixed.

[changelog]
* src/why/why-gen-records.adb (Declare_Wrapper_Conversions):
Take the Is_Moved component into account.
* testsuite/gnatprove/tests/VA05-041__ownership_and_relaxed_init:
Add test.

Change-Id: I7605cfa0db1a00f885b171c45b82eff2746c4d8d