VA25-025 motivating examples for using real eq in functional containers
Currently in the testsuite, most uses of functional containers
contain scalar types for which the logical equality and the Ada one
are the same. File some examples of containers containing strings so
we can see the difference.

* tests/VA25-025__string_map
* tests/VA25-025__string_vector
New tests. Each test contain 2 variants, one with the builtin equality
on strings and the other with a custom equality which also compares the
bounds. Both versions are hard to prove.

Change-Id: I78a0a3324a6379838e96dec0aff5bad93ecde974