V802-025 Fix analysis of body in SPARK_Mode Auto
When the spec of a package has SPARK_Mode Off, either in its public or
private part, the body of the package should not be analyzed at all.
Special care must be taken for expression functions, whose body is
analyzed even in mode Auto, to ensure this is not in such a case.

[changelog]
* src/spark/spark_definition.adb (Mark_Package_Body): Do not analyze
when spec has SPARK_Mode Off. Refine special case for expression
function bodies, so that it only applies to those from with'ed units.
* testsuite/gnatprove/tests/V802-025__spark_mode_off/iomsg.adb
* testsuite/gnatprove/tests/V802-025__spark_mode_off/iomsg.ads
* testsuite/gnatprove/tests/V802-025__spark_mode_off/test.gpr
Add test.
* gnatprove/tests/O416-002__dynamic_predicate_not_in_spark/test.out
* gnatprove/tests/P111-029__marking_delayed_aspect/test.out
Updated error message as expression function body is not analyzed
anymore.
* gnatprove/tests/P712-044__spark_mode/test.out
Remove spurious error, as body should not be analyzed in that case.
* gnatprove/tests/RC17-022__private_pointer/mode_auto.ads
Add Global to expression functions whose body won't be in seen as in
SPARK by flow analysis even if they are available as axioms for
client code ("compatible with SPARK").
* gnatprove/tests/T708-048__flow_analyse_expr_fn/tcp_type.ads
Update test to avoid spurious messages related to expression function
body otherwise not being seen as in SPARK.
* gnatprove/tests/R622-045__pantherchess/test.out
Add new info messages, related to the fact that the expression
function body is not analyzed in Auto mode.

Change-Id: Ic9922f535c82964cb76c7ae60ed8cd2054d945a2