V712-019 Add rule and detection for ghost aliasing through pointers
Aliasing from a non-ghost object to a ghost one should not allow
effect on the non-ghost memory through the ghost (temporary) owner.

[changelog]
* docs/lrm/source/declarations-and-types.rst: Add rule to prevent effect
on non-ghost memory through ghost handle. Renumber rules. No comment or
message in the code of GNAT/GNATprove needed to be shifted.
* src/why/gnat2why-borrow_checker.adb (Check_Assignment): Issue error
message in case of problematic assignment doing a Move or Borrow.
(Check_Parameter_Or_Global): Issue error message in case of problematic
call doing a Move or Free or Borrow.
* testsuite/gnatprove/tests/V712-019__ghost_aliasing
Add test.

Change-Id: I6650c203f962aa4639e00973533d2cf756523f2d