VB29-011 fix checking of limit-subp
When checking if the string given via --limit-subp matches the
location of a subprogram, the previous code returned True for
strings that were textually included in the location, such as
file.ads:12 in file.ads:123. We improve this test by checking that
the substring is either at the end of the string, or is followed by
a ":" sign (which indicates another location in the location chain).

* spark_util-subprograms.adb
(Is_Requested_Subprogram_Or_Task): more thorough checking which
takes into account semantics of sloc string

Change-Id: I7e21793270320e7501fee8e1584236884ac76475